crawlers/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --no-emit-index-url requirements.in
#
apscheduler==3.6.3
    # via python-telegram-bot
async-timeout==4.0.2
    # via redis
beautifulsoup4==4.11.1
    # via -r requirements.in
cachetools==4.2.2
    # via python-telegram-bot
certifi==2021.10.8
    # via
    #   python-telegram-bot
    #   requests
charset-normalizer==2.0.12
    # via requests
commonmark==0.9.1
    # via rich
deprecated==1.2.13
    # via redis
idna==3.3
    # via requests
lxml==4.8.0
    # via -r requirements.in
packaging==21.3
    # via redis
pygments==2.12.0
    # via rich
pyparsing==3.0.8
    # via packaging
python-dotenv==0.20.0
    # via -r requirements.in
python-telegram-bot==13.11
    # via -r requirements.in
pytz==2022.1
    # via
    #   apscheduler
    #   python-telegram-bot
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
pyyaml==6.0
    # via -r requirements.in
redis==4.3.1
    # via -r requirements.in
requests==2.27.1
    # via -r requirements.in
rich==12.4.1
    # via -r requirements.in
six==1.16.0
    # via apscheduler
soupsieve==2.3.2.post1
    # via beautifulsoup4
tornado==6.1
    # via python-telegram-bot
tzdata==2022.1
    # via pytz-deprecation-shim
tzlocal==4.2
    # via apscheduler
urllib3==1.26.9
    # via requests
wrapt==1.14.1
    # via deprecated

# The following packages are considered to be unsafe in a requirements file:
# setuptools
